refactor importing
Move client code to separate file. Make it possible to import some
classes directly from the client module.
