+++ setup.cfg
@@ -1,0 +1,3 @@
+[flake8]
+per-file-ignores =
+    modelon/impact/client/__init__.py:F401
